Semaphore is moved into the sync module.
