+++ glommio/src/lib.rs
@@ -365,8 +365,8 @@ mod multitask;
 mod networking;
 mod pollable;
 pub mod read_write_lock;
-mod semaphore;
 mod shares;
+pub mod sync;
 pub mod timer;
 
 pub use crate::executor::{
@@ -375,7 +375,6 @@ pub use crate::executor::{
 };
 pub use crate::networking::*;
 pub use crate::pollable::Async;
-pub use crate::semaphore::Semaphore;
 pub use crate::shares::{Shares, SharesManager};
 pub use enclose::enclose;
 pub use scopeguard::defer;
